% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVariables.R
\name{extractVariables}
\alias{extractVariables}
\title{Extract variable values for a set of patients and a set of events.}
\usage{
extractVariables(patients, variables, events = NA, snps = NULL,
  transcode = FALSE)
}
\arguments{
\item{patients}{the patients IDs to focus on.}

\item{variables}{the variables to extract.}

\item{events}{the events to focus on. Not taken into account if variables
are not followed-up at the different visits.}

\item{snps}{SNP data to be extracted. Default is NULL ==> no SNP data
are extracted.}

\item{transcode}{a logical indicating if the A/B genotypes should
be converted according to the SNP allele bases (ATCG).}
}
\value{
\describe{
 \item{If several events are provided}{a list of data.frames:
 one data.frame per followed up variable and one data.frame for all
 patient variables. The row names
 of the data frames are the patients IDs (PATNO). The column names of the
 data frames are the events IDs.}
 \item{If only one event is provided}{a data.frame. The row names
 of the data frames are the patients IDs (PATNO). The column names of the
 data frames are the variables.}
}
}
\description{
Extract variable values for a set of patients and a set of events.
}
