% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVarByGroup.R
\name{findVarByGroup}
\alias{findVarByGroup}
\title{Find variables corresponding to a search text in group and sub-group}
\usage{
findVarByGroup(txt)
}
\arguments{
\item{txt}{the text to search in group and sub-group names}
}
\description{
Find variables corresponding to a search text in group and sub-group
}
