% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariable.R
\name{findVariable}
\alias{findVariable}
\title{Find variables corresponding to a search text.}
\usage{
findVariable(txt)
}
\arguments{
\item{txt}{the text to search in variable names and description}
}
\description{
Find variables corresponding to a search text.
}
