% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohort.R
\name{getCohort}
\alias{getCohort}
\title{Get patient IDs of a cohort}
\usage{
getCohort(cohort, enrolledOnly = TRUE)
}
\arguments{
\item{cohort}{the name of the cohort to extract}

\item{enrolledOnly}{if TRUE (default) only patients with an "Enrolled"
status are returned}
}
\description{
Get patient IDs of a cohort
}
