% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MultiPhenoDB.R
\name{getHpByGene}
\alias{getHpByGene}
\title{Get HP by Gene}
\usage{
getHpByGene(...)
}
\arguments{
\item{...}{parameters for the \code{\link{filterGeneByTrait}} function.}
}
\value{
a list providing for each entrez gene ID a set of HP.
}
\description{
Get all entrez HP associated to each gene entrez ID.
These associations come from gene/trait associations
and can be filtered according to origin and clinical signficance.
}
\examples{
##########################
## Filter gene/trait association according to variant origin and clinical status
geneByTrait <- filterGeneByTrait()
head(geneByTrait)
## Get gene/HP association according to gene/trait and trait/HP association
## The filtering options are the same than for filterGeneByTrait().
hpByGene <- getHpByGene()
head(hpByGene)

}
\seealso{
\code{\link{filterGeneByTrait}}, \code{\link{getGeneByHp}}
}

