% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSnpInfo.R
\name{getSnpInfo}
\alias{getSnpInfo}
\title{Get information about PPMI SNPs}
\usage{
getSnpInfo(snps)
}
\arguments{
\item{snps}{a character vector of SNP IDs}
}
\description{
Get information about PPMI SNPs
}
