% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSnps.R
\name{getSnps}
\alias{getSnps}
\title{Get PPMI SNPs data}
\usage{
getSnps(snps, transcode = FALSE)
}
\arguments{
\item{snps}{a character vector of SNPs to be extracted}

\item{transcode}{a logical indicating if the A/B genotypes should
be converted according to the SNP allele bases (ATCG).}
}
\value{
a matrix of character. Columns correspond to selected SNPs.
Rows correspond to patients with a genotype.
}
\description{
Get PPMI SNPs data
}
