% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkRegression.R
\name{plinkRegression}
\alias{plinkRegression}
\title{Call the plink program to test association of a phenotype to SNPs}
\usage{
plinkRegression(
  phenoTable,
  snpSets = NULL,
  mperm = 1000,
  type = c("logistic", "linear"),
  command = "plink --noweb"
)
}
\arguments{
\item{phenoTable}{a data.frame with the first column corresponding
to the endpoint of interest and all the other columns to covariates.}

\item{snpSets}{a list of SNPs sets. If NULL, ignored.}

\item{mperm}{number of permutations to be applied when testing
SNP sets. Ignored if snpSets is NULL.}

\item{type}{the regression type: 'logistic' for cohort comparison
and linear for quantitative trait association}

\item{command}{the path to the plink command. Default is "plink"
but can be changed if needed.}
}
\value{
a list of 2 data.frames: \describe{
 \item{\code{snps}}{Association with enpoint of interest.
 FDR is computed using p.adjust with "BH" method.
 see \url{http://pngu.mgh.harvard.edu/~purcell/plink/anal.shtml#glm}
 for the description of the output}
 \item{\code{cov}}{Association with covariates}
 \item{\code{sets}}{Association at the level of SNP set.
 see \url{http://pngu.mgh.harvard.edu/~purcell/plink/anal.shtml#set}
 for the description of the output.}
 \item{\code{log}}{plink log}
}
}
\description{
Call the plink program to test association of a phenotype to SNPs
}
\seealso{
\url{http://pngu.mgh.harvard.edu/~purcell/plink/}
}
