% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableGroupSummary.R
\name{variableGroupSummary}
\alias{variableGroupSummary}
\title{Summarize the number of variables per Sub-group}
\usage{
variableGroupSummary()
}
\description{
Summarize the number of variables per Sub-group
}
