% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vEvent.R
\name{vEvent}
\alias{vEvent}
\title{Return the visit IDs (EVENT_ID) of a Patient/Visit table}
\usage{
vEvent(d)
}
\arguments{
\item{d}{the table from which EVENT_ID should be extracted}
}
\description{
Return the visit IDs (EVENT_ID) of a Patient/Visit table
}
