% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vPatient.R
\name{vPatient}
\alias{vPatient}
\title{Return the patients IDs (PATNO) of a Patient/Visit table}
\usage{
vPatient(d)
}
\arguments{
\item{d}{the table from which PATNO should be extracted}
}
\description{
Return the patients IDs (PATNO) of a Patient/Visit table
}
