% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohort.R
\name{getCohort}
\alias{getCohort}
\title{Get patient IDs of a cohort}
\usage{
getCohort(cohort, enrolled_or_complete_only = TRUE)
}
\arguments{
\item{cohort}{the name of the cohort to extract}

\item{enrolled_or_complete_only}{if TRUE (default) only patients
with an "Enrolled" or a "Complete" status are returned}
}
\description{
Get patient IDs of a cohort
}
